+++ app/models/image.rb
@@ -2,7 +2,7 @@ class Image < ApplicationRecord
   include Concerns::Visible
 
   include Concerns::Sortable
-  sortable scope: [:asset_id]
+  sortable scope: [:imageable_id]
 
   belongs_to :asset
   belongs_to :imageable, polymorphic: true