+++ app/models/redirect.rb
@@ -1,9 +1,9 @@
 class Redirect < ApplicationRecord
-  validates :source_uri, :destination_uri, :status_code, presence: true
-  validates :source_uri, uniqueness: { case_sensitive: false }
-
   scope :disabled, -> { where(disabled: true) }
   scope :enabled, -> { where('disabled IS NULL or disabled = 0') }
+
+  validates :source_uri, :destination_uri, :status_code, presence: true
+  validates :source_uri, uniqueness: { case_sensitive: false }
 
   def enabled?
     !disabled?