+++ app/models/search_synonym.rb
@@ -1,5 +1,4 @@
 class SearchSynonym < ApplicationRecord
   validates :locale, :term, :synonyms, presence: true
-
   validates :term, uniqueness: { case_sensitive: false, scope: :locale }
 end